SelectMethodMockedTest
- Select Method Mocked Test Added
- Old class Removed
- New Class can be found in Mockito Package
